# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST Skedway",
        "description": "Documentação Skedway.",
        "contact": {
            "email": "bernardodev0809@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/room/create": {
            "post": {
                "tags": [
                    "Sala"
                ],
                "summary": "Criar uma sala",
                "operationId": "d30d1a65610e3c99f848054d0e3345fd",
                "requestBody": {
                    "description": "Dados do sala",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Auditório"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Auditório do primeiro andar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sala criada!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Auditório"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Auditório do primeiro andar"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-29 00:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-29 00:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao criar sala!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Sala não pode ser criada no momento, tente novamente mais tarde!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Erro ao criar sala repetidamente!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Sala existente!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/room/read-all": {
            "get": {
                "tags": [
                    "Sala"
                ],
                "summary": "Ler todas as salas",
                "operationId": "48214369a00dccd2d3b1497383eece35",
                "responses": {
                    "200": {
                        "description": "Salas encontradas!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Auditório"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Auditório do primeiro andar"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-08-29 00:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-08-29 00:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhuma participante encontrado!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Nenhuma sala cadastrada até o momento."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/room/read-by-id/{id}": {
            "get": {
                "tags": [
                    "Sala"
                ],
                "summary": "Ler sala filtrada por id",
                "operationId": "e9b3521900d0ba16cda8c9a86062075a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da sala",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salas encontradas!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Auditório"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Auditório do primeiro andar"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-08-29 00:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-08-29 00:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhuma participante encontrado!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Nenhuma sala cadastrada até o momento."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}